requirements.txt
cmd2==2.4.3
colorama==0.4.6
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
gunicorn==20.1.0
Jinja2==3.1.4
PyMySQL==1.1.1
python-dotenv==1.0.1
setuptools==69.0.3
Werkzeug==3.0.3
postgres
psycopg2-binary
flask_login
flask_wtf
email_validator